implement the ff: - repo - service - controller

+++ src/main/java/org/example/dogapi/dogs/web/dto/CreateDogRequest.java
@@ -1,4 +1,4 @@
 package org.example.dogapi.dogs.web.dto;
 
-public record CreateDogRequest(String name, String breed, int age) {
+public record CreateDogRequest(String name, String breed, short age) {
 }
