+++ src/main/java/org/example/dogapi/dogs/web/dto/UpdateDogRequest.java
@@ -1,4 +1,4 @@
 package org.example.dogapi.dogs.web.dto;
 
-public record UpdateDogRequest(String name, String breed, int age) {
+public record UpdateDogRequest(String name, String breed, short age) {
 }
